'stateForDelete'

+++ api/controllers/Organize/Courses.controller.js
@@ -151,7 +151,7 @@ const editCourse = async (req, res) => {
     if (courseid.State !== 'pending') {
       return res.json({
         msg: 'Your course is approved, you can not edit it',
-        statusCode: '1',
+        statusCode: errorCodes.approvedCourse,
       })
     }
     CoursesModel.update(Course, { where: { id: courseID, accountId } })
@@ -183,6 +183,12 @@ const deleteCourse = async (req, res) => {
       return res.json({
         msg: 'course doesnt exist',
         statusCode: errorCodes.cousrseDoesntExist,
+      })
+    }
+    if (courseid.State !== 'pending') {
+      return res.json({
+        msg: 'Your course is approved, you can not delete it',
+        statusCode: errorCodes.approvedCourse,
       })
     }
     CoursesModel.destroy({ where: { id: courseID, accountId } })
